Document DocMind entity-first persistence shift
