+++ docs/chunks/S13-DocMind-RefactoringPlan.md
@@ -2,6 +2,7 @@
 
 ## Current Coverage Snapshot
 - **Durable pipeline core in place.** The hosted `DocumentProcessingWorker` advances `DocumentProcessingJob` ledgers through extraction, embeddings, vision, insights, and aggregation while emitting persisted telemetry, delivering the resilient orchestration envisioned in the proposal’s AI processing plan.【F:docs/chunks/S13-DocMind/03_ai_processing.md†L35-L66】【F:samples/S13.DocMind/Infrastructure/DocumentProcessingWorker.cs†L17-L418】【F:samples/S13.DocMind/Models/DocumentProcessingJob.cs†L15-L170】
+- **Entity-first persistence enforced.** Repository helpers were removed in favour of `Entity<T>` access directly inside intake, diagnostics, discovery, and worker flows, ensuring projections, queue slices, and retries operate through the canonical static APIs recommended by the proposal.【F:samples/S13.DocMind/Services/DocumentIntakeService.cs†L78-L205】【F:samples/S13.DocMind/Services/DocumentDiscoveryProjectionBuilder.cs†L20-L214】【F:samples/S13.DocMind/Services/ProcessingDiagnosticsService.cs†L52-L236】【F:samples/S13.DocMind/Infrastructure/DocumentProcessingWorker.cs†L39-L139】
 - **Operator visibility partially delivered.** Discovery refresh scheduling, vector bootstrap checks, and diagnostics endpoints expose queue slices, vector readiness, and replay scaffolding, giving operators insight into the live system but without the incremental and contract guardrails the proposal targets.【F:samples/S13.DocMind/Infrastructure/DocumentDiscoveryRefreshService.cs†L12-L173】【F:samples/S13.DocMind/Infrastructure/DocMindVectorHealth.cs†L8-L109】【F:samples/S13.DocMind/Services/ProcessingDiagnosticsService.cs†L15-L210】
 
 ## Delta Analysis
